Use pagination for YouTube API requests
So, this is a fun one.

YouTube does not return all comments all at once for a livestreaming
video. Instead, it only returns X most recent comments. So, this means
we have to paginate through the API to get more comments.

However, there are serious performance concerns here. YouTube mandates
that you sleep for X seconds before making another API request (and THEY
set the sleep time), so this can significantly lengthen the time for a
response from the API. To combat this, we will only get the last 5 pages
of comments from YouTube. ¯\_(ツ)_/¯ IMHO, this should still give enough data to get a
good sense of what's going on in the live chat.